refactor: Modify reducer name
ChangeGradeReducer에서 ChangeMemberInfoReducer로 변경했습니다.
회원 삭제 기능을 넣기 위해서 이름을 바꾸는 것이 좋다고 생각했습니다.
