+++ frontend/src/_actions/changeMemberInfoAction.js
@@ -1,4 +1,4 @@
-import { CHANGE_GRADE, INIT_MEMBER } from './types';
+import { CHANGE_GRADE, INIT_MEMBER, DELETE_MEMBER } from './types';
 
 export function changeGrade(data) {
   return {
@@ -10,6 +10,13 @@ export function changeGrade(data) {
 export function initMember(data) {
   return {
     type: INIT_MEMBER,
+    payload: data,
+  };
+}
+
+export function deleteMember(data) {
+  return {
+    type: DELETE_MEMBER,
     payload: data,
   };
 }
