+++ frontend/src/_actions/types.js
@@ -1,3 +1,4 @@
 export const REGISTER_USER = 'REGISTER_USER';
 export const INIT_MEMBER = 'INIT_MEMBER';
 export const CHANGE_GRADE = 'CHANGE_GRADE';
+export const DELETE_MEMBER = 'DELETE_MEMBER';
